Figuring out timer function
For the targets mode I intend to have a timer to measure how long the game goes for. The timer function is something I have not sorted out yet, and so far I have been experimenting with coroutines and measuring framerate to find realtime. I will keep looking into this until I find a proper solution. Stopping the game when the timer runs out is simple enough though, only needing a conditional statement to check whether the value is at or below zero.