Target Spawning Mechanic Finished
The base mechanic for this gamemode is done. On hit, targets are teleported in a random position in a radius. Scoreboard and timer work as well, on timer hitting zero the game stops and waits for the target to be hit before starting to count the score again. This is all tested without the proper foil and vr controller, but in theory with the gameObject tags there should be no problems. At this point in time I will be attempting to polish the score sort on the board, and polish UI. Eventually I will also consider using a local or global database, but at the moment that is beyond the scope of this gamemode.