Target Almost polished
UI and scoreboard have been implemented, but will still need lots of refinement. I have successfully imported a public domain font for the scoreboard, and have chosen to use an insertion sort to sort the top 3 highest scores on the scoreboard. However, it does not properly work and will seemingly replace scores at the wrong place. This may be due to the values being inputted on the go instead of all at once, but in theory that shouldn't be a problem. Once the UI has been improved and the scoreboard is properly handled, then this gamemode will be done! Then I will be able to test with the VR controllers and see if everything works properly.