Further Foil Testing
I recently discovered that while using the transform.position values of Unity, it does not use the Unity Physics engine, and therefore objects moved by transform do not have any velocity or force. This realisation is very important to making the foil bend, as it should be bending with a certain threshold of force even when not colliding with an object. Certainly, it would be easier to test the foil without usage of another collider. From here, I will be attempting to attach a separate GameObject that is the foil onto a hand in a separate scene, and apply the theoretical velocity onto the rigidbody velocity through the physics equation v = s/t. An additional note, while this is on the TargetTraining branch, it is not actually going to affect the targets very much. Henceforth I will be using the FoilPhysics branches for these experiments.