Refining Targets Mode
To note: this was the first time developing the project on different hardware, in which a HTC Vive was used instead of a HTC Vive Pro. VR controls were a complete success, guaranteeing what was once theoretical probability on the program running on all headsets with SteamVR support.
The objective of this push was to refine the target mode, find any flaws and try to polish the scene and scripts as much as possible. TargetBehaviours script is modified and more modularised to clean up the Update() function, and the VR controller was imported to test how the mode would fair with a foil. As the foil is still incomplete, bending the foil for a flick was still not an option.
Still to be resolved is the sorting algorithm for the scores and the cleaning of hitting the target. Currently, the foil will sometimes be able to hit the target multiple times before the target properly teleports away, and the foil hitting the target in a position where the target spawns on top of the foil is an issue. To resolve, a cooldown can be implemented within the target, ensuring that the user cannot hit the target multiple times with one thrust. Additionally, having a system where the target only registers a hit if the foil is over a certain velocity will also be implemented, but this requires the foil to be a separate GameObject. This issue can be resolved after the foil is refined more, hopefully even with proper physics for the flick. Overall this push was successful to a degree, but there is a clear path to resolve all the issues for the Targets Gamemode.