Beginning of AI Development - Animating the Model without VR Hardware
The objective of this commit was to:
- Import the model that will be used for the AI
- Import the AI Plugin correctly into the Unity Project
- Ensure the model is ready to be controlled by the computer

The first two objectives were achieved with ease. Importing the plugin and model were fairly easy, just by using Unity's Package Manager. However, the third objective has yet to be completed. This is due to animations for the model. Currently I am working towards making the animations play when a certain condition is met, e.g. moving forward. However this is not working as expected, as the conditional triggers seem to not be activating. Further debugging is required, however the nature of the animator variables mean that output statements are much harder to manage, however that seems like the easiest way to see if the variable is affected and/or if the condition is being properly met.
Overall, this is a good start to this branch, however much more work and research will be necessary for a polished final product.