Model Animation Complete
The objective of this commit was to be able to conditionally trigger animation on the AI model without use of VR Hardware. This commit is a success.
A new script was written to translate the object across the world space, which proved to be a more effective testing measure than previously moving the model manually during play. Going to the animator tree as well showed the bug within the system, that being with the conditions of which the model moved to were incorrect.
At first, when the model moved, the animation would be the wrong walking direction, which was fixed by reversing the previous threshold of the spectrum. Then tweaking the condition to be more sensitive to movement proved to be effective in ensuring the model would move at slight movements.
Overall, this commit was a great success. Further development now will be concentrated on the AI controller, and using the Candice AI unity plugin to develop the AI.