[WIP] Development of Fencing Behaviour using Candice AI
While this push is not necessarily complete, it is important to frequently update the branch for logging the development progress, especially for this specific area where there will need to be lots of editing.
The purpose of this push was to develop a default behaviour using candice AI. While there is a plugin that assists in this regard, using a visual development UI, it is a business expenditure deemed unnecessary.
What an AI behaviour will do is give the AI a way to interact with its' environment without external input. The goal for this behaviour so far is just to attempt to move to a position where a constant distance is maintained between itself and the player model. As the animation for the model is triggered by movement, animation triggers are unnecessary for this AI behaviour.
In particular the way I have discovered to add and develop behaviours is in the CandiceAIController file, along with the BehaviorTreeMono file. Adding an option and a series of CandiceAI variables and objects have allowed me to construct a proper template, and add a method for the behavior. After finding the direction and distance to maintain, the model should call on a movement method that was already included within the file.
A problem encountered in this push is a runtime error. Whenever the "Fencing" AI behaviour is selected and program executed, the solution will crash Unity. This will be a major issue to be resolved in hopefully the next push or two.
Another issue to consider is the lack of documentation for coding using CandiceAI. For this reason further research will be conducted, both for documentation for CandiceAI and for alternate methods of AI development in Unity.
Overall, this push has been a mixed success. While I am learning how to use CandiceAI, the lack of general support coupled with a major runtime error is a major deterrant.