[WIP2] Development of Ai
The approach between the last push and this one has changed drastically. On the understanding that the AI controlling the fencer will be simplistic, the decision to write my own algorithm over using the potentially convoluted plugin designed for more complex AI has been made.
The objective of this push was to generate an algorithm for the fencing AI that would keep within the specified bounds of the arena yet as parallel to the player as possible, and would attempt to maintain the same distance between the player and AI. Other modules have also been integrated to ensure that the more advanced features can be implemented with ease.
After much research on AI algorithms (going through possible other plugins and the A* pathfinding algorithm) I have decided that this AI is enough to be written in an algorithm.
The parameters were that the AI would attempt to maintain distance, and in the scenario where the player would suddenly move it would attack or attempt to parry if the player was close enough, otherwise it would chase the player until it's in an adequate distance. This push aims to ensure the AI can stick within the boundaries of the arena and parallel to the player.
This feature is still under development. As there are many ways to determine whether a vector is parallel, I have been experimenting with which way is optimal, and have landed on the function Angle, which can safely determine whether the angle between vectors is within the threshold of either 0 or 180. At the moment, this is still being developed and has not yet been completed.
Additionally, some files have been moved to folders for organisation.
Overall, there is still much work to be done, and for this reason this push is not a complete success. However there is a clear pathway to this feature, which will prove very effective for the development of said feature.