Maintaining Distance - Development of AI
This push aimed to create a script that would allow the AI to maintain a certain distance away from the player.
Many different methods were experimented with. This ranged from using cross/dot product to determine parallel nature, and calculating distance to walking direction. Eventually a more suitable solution was found: Vector3.MoveTowards()
With this inbuilt method all that had to be done is make the target the player's position with a slightly adjusted x component to be in front of the player. The maintain distance function is almost complete, however a suitable formula for walkspeed is yet to be completed. The aim is for the speed to increase the most when the player is either very close or very far. This means that the function for walk speed should be a parabola, but I just haven't found the correct equation yet. Alternatively I could use an if statement but it may be more practical to use one formula.
Overall, this push was a success.
Lots of work and efficient modularisation has been employed, and the hardest function for this module has been completed.
Also to note is that the Candice AI package has been deleted to save file space.