Maintaining Distance V.2  - Development of AI
This push aimed to amend discovered issues with the movement of the AI. Two issues arose from the previous push: firstly, the walking speed was not up to standard when moving along the Z axis, and secondly, the AI would have no way to recover being pushed out of the arena, and theoretically the player could just move forward until the AI is no longer within the bounds of the bout.
Tackling walk speed first, the original intention was to derive a formula that was parabolic or hyperbolic in nature, where the walkspeed would be high if the player was either close or far. This method took lots of mathematics to perform for a subpar result. The most refined formula was y = 2(x - z)^2, where y is the walkSpeed, x is the distanceFromPlayer, and z is the intended distance to be maintained. Mostly, the issue arose from close range, where the AI did not have a walkspeed to keep up with the player. The alternate integrated method was using 2 separate linear formulas depending on whether the player had crossed the distanceToMaintain threshold. This used just a regular linear formula. However, this didn't adequately cover the walking speed either. Revisiting the parabolic formula,  one of the main issues was overcome by rounding the values to the nearest integer. This successfully solved the issue with close proximity speed, but another issue was found with the speed when moving along the Z axis (side to side). This was overcome by always setting the minimum walking speed to 1 instead of 0. This adequately fixed this issue, and the module for maintaining distance was complete.
The bout boundary problem was more of a focus for this push. Originally, what was integrated was just stepping out of bounds would cause the AI to stop moving. In future there would be another script that would calculate score based upon this factor among others. However, the player would be able to run the opponent out, which would be a very suboptimal for the program. To solve this, another boundary collision box was added, this time smaller and not out of bounds. This would ensure that if the AI was physically pushed out then it would return the point to the player, but the AI would resist to going outside of it's boundaries. Originally, the OnTriggerStay and OnTriggerExit were employed for both functions, however, an issue that was found was the incorrect processing of collisions within these statements. It would not properly run through one of the conditional statements. To put less stress on the module, I instead used raycasts for some of the processing. Using a layer mask and a downwards raycast to detect whether the AI was still within the bout was most appropriate, and the OnTriggerExit functioned worked as expected. Another raycast was implemented to detect behind the AI model whenever it would go outside it's own restrictions, detecting if the newly set collider would be hit. This signalled the back of the arena, and would tell the model to return to the middle and advance forward. Elsewise it would just return to the middle of the arena.
While both solutions were found, a major issue at the moment is the processing of animations. As there is no side step animations for the model in the prefab, the model looks slightly unnatural moving side to side, shaking when it uses the walking forward or backward animation. Further testing and perhaps blender animation is required to fully fix this issue.
At the moment the plan is to begin the attack conditions and sequences. I will have to fiddle with more animations, so given the chance the previous problem will be fixed. The inclusion of the foil for the AI may also be necessary, but at the moment development on the physics of the foil is incomplete, and will be integrated at a later date.
Overall, while there is still a sizeable amount of work to do, this push was a success and greatly propelled this gamemode forward in development.