Attacking and Lunge & Optimising Moving Back into Bout - Development of AI
The objective of this push was to successfully implement the use of chasing into attack and lunging functions of the AI. This mainly involved implementing a theoretical solution that works and has been tested.
Using parameters for the range needed to chase or lunge,  conditions would be checked within the FixedUpdate() module. If the chase condition is met, the model would rush towards the player. This simply updated the transform position over time using Time.deltaTime, and works flawlessly. Upon crossing the threshold to lunge, the animator variable would be set to true, and a boolean which disables other functions would be set true. An event is called upon once the lunge animation is complete, unlocking other modules and allowing the capacity for movement. A timer will also start on a separate thread that prohibits multiple attacks within a short period of time. All modules work as intended.
The move back function was also optimised to work under more conditions. An issue that arose was the model swapping positions too quickly between the back or side of the bout and the player. To solve this issue, a float variable that would be counted down with Time.deltaTime. The raycast will then only be called once the timer had reset. This fixed the problem and allowed the model to move back into the bout properly.
Overall this push was a success, with no created issues and two subroutines successfully implemented. Next will be UI, and once the foil physics are complete importing the VR model in and working on the Parry subroutine.