AI Mode UI Development
The objective of this push was to fully develop the UI as well as possible on this machine, which does not have access to VR. This push was a mixed success.
Starting, the stopwatch and Bout timer were to be added. This was simply done mostly within the Unity scene editor. The UI Canvas was set up to the correct proportions. Then TextMeshPro was imported to further help with UI text objects. This plugin is widely accepted by the Unity community as the superior method of text formatting. The text was formatted and placed on the canvas, and the colour palette was picked. Using coolors.co to generate a palette based on existing colours on the scene, multiple font presets were made to accommodate for different coloured text. After the formatting of the timer and bout counter, I conducted some research on using a curved mesh for the main scene menu. Experimentation with blender was conducted and a simple two-sided curved mesh was generated, with corresponding materials created for the colour. However, the buttons will have to be developed in VR, and as of now I do not have a set way to format selections in VR. These will need to be developed with a VR headset from here.
Overall however, this push was a success, as the beginning of the UI has been successfully generated.
Once development with VR has started, the parry function will be developed along with the rest of the UI.