Post Processing Test for UI
The aim of this push was to begin to produce details on the UI using the Universal Render Pipeline (URP) and post processing to achieve a stylistic menu. Overall, this push was a failure and this branch will shortly after be closed, as functionality is the main goal.
Originally garnering inspiration from old CRT screens, the goal was intended to add a post processing effect on UI menus. The initial idea was to develop the shader from scratch, and then apply said shader to only UI menus. Implementing the shader was easy enough with a guide, and a script was applied to create a new material that has the shader information. The CRT screen was working ok, but the effect was applied to the full screen. After some more  research, the post processing effect works by being applied to the final image produced by the camera. By stacking cameras, it was possible to render some game objects after the post processing has been applied. This method seemed to work, but wasn't able to solve the issue as the post processing effect was still applied to the entire screen.
After this, I tried another approach. Using the URP plugin, more options were created to control graphical elements. More advanced options were made for camera stacking, but that method still continued to not work. The next method was to create a render texture and shader graph, and make the camera render a canvas that applied the effect. However, this was limited by VR, as a fullscreen canvas is not attainable to the VR headset. After about a week of testing, the decision was made to continue working on the UI elements instead over the post processing.
While this push was a failure, I was able to learn more about the URP and more on graphics. I have also gotten more information regarding effective UI creation. From here the UI for the target scene will be developed. Once the foil has been fully completed, this branch will merch with another for a more holistic approach to UI design.