Bezier Curve Generation for Foil Bends
The objective of this push was to develop bezier curve generation in a 3D space, theorizing on the application of this on a foil. Overall, this push was a success.
This stemmed off of a video found here: https://www.youtube.com/watch?v=w_qdI6daqyo
The fishing rod bending seemed to be something that could be applied to the foil. Looking in the comments some people asked how the fishing rod bend was achieved, in which the creator of the video said overall, they used "Catenary Equations", "Bezier curves" and "Juan Carlos bones". Catenary is the term specific to ropes and chains that are suspended between two points, where bezier curves are lines and curves generated in computer graphics through a specific equation. Carlos' bones were more elusive to find a definition on, but I deemed them not necessary. After researching more, an indepth tutorial on linear and quadratic bezier curves were found in the following links:
https://www.youtube.com/watch?v=Xwj8_z9OrFw
https://www.youtube.com/watch?v=wgCxcOXMGgA
I modified the algorithms slightly and added some more components that may be necessary for the foil. The algorithm generates the correct positions of the subsections, which can be applied to the bones of the foil. The midpoint can also be calculated using standard mathematics for each component (m = x2+x1/2). However, the value of the midpoint must be altered to be in the opposite direction of the foil swing, which I have not figured out yet. The velocity threshold to bend has also not been explored fully, but that can be possibly derived by altering the velocity value and multiplying it to one of the values of the midpoint.
Overall this push has been a success and a bit more research is required until the foil bend is fully complete.