Foil Flicks and Bends Development
This push stores changes from the past couple of months due disorganisation.
Explored in this push was chain link-like physics, using joints and script-generated objects. While this method worked as intended (being a chain), there didn't seem to be many configurable options allowing rigidity or movement without the physics engine. Also explored were bezier curves, in two ways: importing from blender as a path or attempting to utilise an animation curve. The blender model import option was used first, however this proved futile as bezier curves are not supported when unity imports .fbx files, and the only way to import curves is through making them a mesh, thereby creating a static object, which did not work.
While overall not a success, further research is conducted to find a suitable option for the foil to bend.