Targets Polishing + UI
The objective of this push was to start finalizing the targets gamemode, fixing any potential bugs and issues stated in the Alpha test 1 and implementing UI.
Beginning with UI, Text Mesh Pro (TMP) was installed and replaced the previous Unity text meshes. TMP is universally more acceptable, allowing extended control of UI elements. TMP is also supported by Unity, allowing the transition between the two to be seemless. A colour palette was also generated using coolors.co and colours sampled from the sports hall, allowing some more expansive colours. The scoreboard menu was finished, and is a complete success. XR Interaction menu is currently in development now.
In regards to bugs and other issues, most have been fixed or have been planned to be fixed for a while. The target's hitbox has been fixed and changed from a capsule collider to a mesh collider. It has also been changed to trigger, along with the method for detecting hits. Previously, OnCollisionEnter was used, meaning that when two hitboxes collided, the method would be called. This lead to issues arising when the foil became stuck inside the target, hitting multiple times within a frame. Changing the hitboxes to triggers and using the OnTriggerEnter method rectified this issue. Additionally, the conditions for increasing score now requires the user to hit the target at a certain velocity, to ensure actual improvement reflected in the score. Later, once the foil bend has been fully implemented, the tip will be the only part of the foil to increase the score too.
Audio will also be a consideration in the coming pushes. While not urgent, this will greatly increase the user experience while playing.
Overall, this push was a success, with much more planned ahead.
(Note as well that the bezier curve script seemed to have been copied to this branch from the foil testing branch. This will not affect development in any way.)