Bezier Curve Generation - Applying the Line Renderer Function to an Array of GameObjects
The aim of this push was to apply the same logic and similar code from the line renderer to an array of gameobjects. This push was a mixed success, but a substantial leap in progress.
A new script was created for the purpose of recycling some logic while applying it to bones. The bezier curve generates properly, but the issue now lies within creating a smooth transition to the bend and creating conditions for the bend to occur. Some issues that were occurring was improper position placement, incrementing out or not far enough within an array, and magnitude of the bend being unsuitable. Most of these issues were resolved with small adjustments, such as adjusting the value of incrementers.
Issues that still need to be resolved is the use of smoothing with the control point, the control point being affected when moved along the foil, and conditions for the foil to bend. These issues are still being worked towards, but an alternate solution to a constantly updating multiplier based on the velocity may be required.
Overall, however, the bezier curve generation now works to a degree on an array of gameObjects functioning as bones of a model. This push was a mixed success, and the implementation of this feature will not take too much longer.