Foil Bend Conditions and Bezier Curve
The objective of this push was to apply smoothing to the bezier curve, and apply conditions for the foil to bend properly. This push was a mixed success.
Smoothing was achieved with relative ease using the Lerp (Linear interpolation) function on vectors. This allowed a variable to control the speed at which the control point approached it's final destination, allowing the bend to apply smoothly to the foil. In this regard, the push was a success.
However, the conditional aspect of the flick is a major issue at the moment. A video that was filmed and uploaded to the issues tab on the repo demonstrates how the bend is caused by the inertia of the tip, essentially showing that the tip moving is what is causing the bend. Previously, attempts were made to try make the foil bend around the location of the tip based on the difference in velocities of the tip and base, but this is no longer an option as the tip has to be the one that is moved. As a result the controlpoint also will not have to be adjusted as much.
Further exploration will be made with sudden changes in acceleration, most likely deriving the velocity equation v = u + at to calculate accleration, and cause the tip to move accordingly.
Overall, this push was somewhat of a mixed success, but the current issue is very large.