Foil Bend - Spring Approach
The objective of this push was to develop conditions for the foil to bend, or alternatively, find another way to make the foil bend. This is an overall success.
A comment made on the Foil Bend Issue on the repo gives a brief explanation of the process of swapping methods. Essentially, the previous method attempted to contract the tip after a specific condition is met that would cause the foil to bend. It would also alter how far the tip travelled during the bend. This method had many flaws, requiring complex physics and mathematics to function at what was predicted to be subpar. Research on possible alternative methods were explored.
Another method explored was using spring, hinge and config joint components in the Unity scene. These components were pre built, and relied on the Unity physics engine more. However, the joints did not function as expected, and with more research the joints were declared to not function in a way that would benefit the foil.
(The following can be seen in the script: "springtest.cs"). This idea of using joints and springs however lead to an idea for an alternate method; creating a configurable spring joint in code that was affected by movement. Derived from this thread (forum.unity.com/threads/best-method-to-make-rigidbody-spring-to-upright-rotation.1126985/), the chain of gameobjects required an empty gameobject acting as an anchor point for rotation between each physical object. This approach has worked the best by far, and this approach can be adjusted based upon the flexibility of the bones to move in the worldspace, the spring strength and the damper strength.
Further testing in the actual foil gameobject and in a VR environment is required, however, this push has been overall a great success.