Ray Interactors, Menu and UI Art Style complete
The objective of this push was to properly integrate the raycast menu system and to finish the design of the UI. This push has been a complete success.
With the UI from previous push complete, the ray cast system was all that needed to be fixed. Using Valem's guide to making a VR game [INSERT LINK], the raycast system was completed. The rays were then adjusted to be transparent when not on the menu. When the branches are merged, a foil object will appear when not pointing towards the menu.
This did bring up an issue concerning the size UI menu's. Further research reveals that Unity units are measured in metres, so the menu was adjusted by a factor or 1/2 from it's previous iteration, which fixed the sizing issue.
Next will be the development of settings, which will include audio levels, height calibration and dominant hand switching.
Overall this push was a large success and has given a template for further UI design.