Target Mode Change
This push is marks how the Target gamemode is necessary to change.
During testing and merging the physic and target branch, many bugs still persisted in the game. This was caused by collision, where the score or teleport of the target would not work properly. More bugs detailed on the issues page on Github (https://github.com/AideenMonike/f3nce/issues/11).
This caused me to rethink how the targets gamemode was supposed to function. As an application to improve accuracy and reaction time, this gamemode did neither, instead encouraging the player to rapidly wave the foil to hit the target as many times as possible.
An alternative that will be explored is measuring the accuracy and reaction time to hit the target. This mode will be measured by number of targets hit, recording misses and time to hit. It will then calculate the average of reaction times and record misses to showcase the score.
Lots of logic from the previous mode will be utilised, but the script does need to change.
While overall a daunting task, this push signifies new progress to the project.