Redoing Targets Gamemode - Functionality
This push aimed to redevelop the target gamemode with an enhanced focus on accuracy of attacks and reaction times of the attacks. This push was a success, developing the functionality of the targets with some UI elements added.
The original plan at the beginning of the development of this push was to have a coroutine/multiple coroutines running at once, which would handle different functions. At the very least, coroutines would have to be used in order to properly integrate timer functions and other time - sensitive methods. However, some coroutines were structured incorrectly, and the "yield return" statements would not properly function thereby pausing the coroutine at a loop.
This was fixed by only using one coroutine, and instead allowing the coroutine to run through a loop determined by how many targets the player requests. The coroutine would call on numerous other methods.
In particular, the coroutine would pause and measure the time taken to hit the target with a while(true) loop. Using a "yield return null", the time between frames would be measured and added on to a variable.
A scoreboard of target times for one round was also implemented, listing the times to hit the target. The scoreboard has been placed on the side, and rotates towards the head position of the player.
A useful feature to note is that sprites will load properly with large width and heights, but not smaller sizes. This can be circumvented by using the local scale of the canvas, allowing more minute movements and loading of sprites on the menu.
Overall, this push was a success. From here, an algorithm to determine the score based on misses and reaction times will be made and record the top 3 local highest scores in the menu.