Target Gamemode Enhancements - Misses, Reset Position
The objective of the push was to enhance the targets gamemode as much as possible, including some last small features and quality of life before fully integrating UI and audio. The overall push was a success, implementing two new features: Misses and Returning to base before the next target appears.
In terms of counting misses, child objects were added to the target; a mesh and box collider. This functioned to distinguish between colliders. The original approach to this issue was to create an array that would be populated with colliders, and then the ProcessCollision() method would decide a different outcome based on the colliders. However, there was trouble distinguishing between colliders, so another approach was used. Using the Unity event system, methods could be called from other scripts when a specific condition was met. A new script was added, which included all OnTrigger methods that would invoke the events. The events would then be assigned to the target gameobject and would then trigger the function within TargetReact.cs. To do this from the inspector however, the event would have to be serializable, and Unity 2018 does not support generic event classes. So another script was added which contains system serializable event classes.
A problem that still needs to be fixed is the collider not registering a straight hit. Currently, since the foil triggers both colliders, both methods are called. This problem is minor, and can be fixed in due time.
The other feature was resetting the position of the fencer. While no locomotion system is or will be present in the targets gamemode, this feature prevents the player from straying too fair from the origin and hitting targets without proper spacing techniques. Now, if the player strays from the original position, the target will wait for the player to return to the origin before appearing again. This was implemented with the previous collision detector that was used for missed targets, this time instead setting a boolean to true or false. This worked without any problems.
Overall, the push was a success. The targets gamemode is close to complete, and UI is the main priority from here.