Cleaning up foil bending
This push aimed to get the foil bending working properly and the sandbox gamemode ready for it's first alpha test. This push was a success.
The task with the foil was to polish any collider bugs and collision tasks, which was a staggering success. One of the problems encountered was the colliders on the foil hitting each other and not bending properly. While the main cause for the issue was not found, the issue was mitigated by shifting the colliders onto the joints. It was also strange how the foil interacted even with no contact points.
The other problem was colliding with other objects. When going to deep into another collider, the joints would scatter and end up malformed when taken out of the collider. One approach to this was to dynamically/manually set the position once the colliders exited. However, this still resulted in some minor deformities, so freezing changing positions on the rigidbody component of each joint solved the issue.
The spring, damper and flexibility values were changed also to suit the new colliders on the joints.
Finally, a model with animations was added to the scene along with a couple objects that can be interacted with. Planned in development is swapping animations with UI and more objects to interact with, along with scene resets.
Overall, this push was a success.