Post Alpha Test 2 & 3 Adjustments
This push aims to solve some minor adjustments after the Alpha Tests for Targets and Sandbox.
Things that were fixed during the test were:
- Target teleport range
- Behind line collider
- Plane collider for sandbox
Following this, some other minor adjustments to be solved are:
- Colliders for the sandbox dummy
Overall, the alpha tests were a success, and more detailed feedback is soon to be given.