Targets UI and Minor Adjustments
This push aimed to develop the UI of the targets mode more and adjust any preexisting issues. This push was a major success.
The first issue to be resolved was the implementation of a UI menu that adjusts the target size and number of targets per round. Using the prior knowledge gained in the home menu, this was easily achieved. Values were also integrated into the targets script well, using the in-built events that the UI elements had. As a result, the menu values being changed in VR was successful. However, a help button is planned on being developed for demonstration and tutorial purposes for the user.
Another issue that was noticed during alpha testing was an inclination to not notice when a round is over. To counteract this, a menu is drafted to appear when a round is complete, and 5 seconds passes before the target resets. While the UI menu is not completed, the delay is successfully integrated.
Keeping the player behind the line was also further developed, using a UI text element to alert the player to return. The target will also not appear until the condition is met. Originally, the text was supposed to track around the player's head, however, a logic error caused the text to instead not stray far from one point in front of the player, which was deemed more suitable and stylistic.
In addition, a scoring algorithm was added which calculates score based upon the average times, the number of misses and the size of the target. This function had some issues with the UI menu array length changes, however this issue was rectified by only changing the array length when the game isn't started.
Overall this push was a success, and requires only a couple more features such as the round complete screen and overarching game features (height calibration, dominant hand swapping, etc.)