Sandbox Menus and Dummy Animations
The objective of this push was to create appropriate menus for the sandbox gamemode, and assign colliders and animations for the dummy. This push was a success, but some more work on the menus need to be done.
Two menus were developed; one for dummy animation and resetting the level, and one for giving help on controls and less useful level options, such as spawning more objects and changing the map scene. Creating the menus for XR was fairly simple, using the same methods used for the Home Menu. XR ray interactors were placed on the controllers, and a tracked device component was added to the UI canvases. With the smaller menu, the canvas was placed as a child object of the main camera as to always track in front of the player, with a button on the controller toggling it's visibility state. An issue that arose was the different headset controllers requiring different bindings, with oculus requiring the menuButton binding and the Vive wands requiring the primary binding. To overcome this issue multiple bindings have been set for the same input function. The personal menu is mostly operational, with the SceneManager resetting the level and the animation for the dummy functional also. Once merged with the Home Menu branch after the next couple of Alpha tests, this menu will be completed fully to return to the main menu.
Additionally, the animations for the dummy have been complete and the object fitted with multiple primitive colliders. Changing the parameters on the animator to booleans instead of triggers, and animations fitted with longer transitions and exit time to appropriately fit with the menu options. This worked flawlessly.
Overall the push is a success, but the larger stationary menu needs to have subroutines fitted to it's buttons and some more objects will be placed.